+++ examples/hello_world.rs
@@ -16,7 +16,7 @@ fn main() {
     error!("you can see this error log with stack trace");
 
     // remove logger's stack plugin
-    logkit::default_logger().write().unmount("stack");
+    logkit::default_logger().write().unmount(|t| t.as_any().downcast_ref::<logkit::StackPlugin>().is_some());
 
     error!("stack trace printing feature has been disabled");