+++ src/define.rs
@@ -1,9 +1,8 @@
 //! Log levels and Encode trait
+pub(crate) use std::any::Any;
 pub(crate) use std::io::Write;
 pub(crate) use std::path::Path;
-pub(crate) use std::borrow::Cow;
 pub(crate) use std::cell::RefCell;
-pub(crate) use indexmap::IndexMap;
 pub(crate) use parking_lot::Mutex;
 pub(crate) use parking_lot::RwLock;